refactor(minarch): extract runframe/audio callbacks, prune dead includes
- Extract chooseSyncRef, limitFF, run_frame → minarch_runframe.c/.h
- Extract onAudioSinkChanged, Audio_checkAndResetIfNeeded → minarch_audio.c/.h
  (renamed to Audio_onSinkChanged for consistent module prefix)
- Remove ~25 includes from minarch.c now dead after prior extractions:
  system headers moved with their callers (fcntl, dlfcn, libgen, time,
  errno, zip, glob, lz4, HAS_SRM streams, dirent, rcheevos) and headers
  already transitively available via minarch_internal.h (SDL, unistd,
  sys/stat, pthread, libretro, defines, api, config, utils, scaler)
