+++ workspace/all/minarch/minarch_audio.h
@@ -4,3 +4,5 @@
 
 void audio_sample_callback(int16_t left, int16_t right);
 size_t audio_sample_batch_callback(const int16_t *data, size_t frames);
+void Audio_onSinkChanged(int device, int watch_event);
+void Audio_checkAndResetIfNeeded(void);
