+++ workspace/all/minarch/minarch_runframe.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void chooseSyncRef(void);
+void run_frame(void);
